signed / unsigned updates

+++ libs/src/str/strutil.c
@@ -4,9 +4,9 @@
  * and conditions for redistribution.
  */
 
+#include "config.h"
 #include <string.h>
 #include <ctype.h>
-#include "config.h"
 
 #ifndef NULL
 #define NULL				0
@@ -82,8 +82,8 @@ char *str_casefind( register char *str, char *sstr )
 char *str_casefind( const char *s1, const char *s2 )
 {
    int i ;
-   int l1 = strlen( s1 ) ;
-   int l2 = strlen( s2 ) ;
+   unsigned long l1 = strlen( s1 ) ;
+   unsigned long l2 = strlen( s2 ) ;
 
    for ( i = 0 ; i < l1 - l2 + 1 ; i++ )
       if ( strncasecmp( &s1[ i ], s2, l2 ) == 0 )
@@ -101,3 +101,4 @@ void str_fill( register char *s, register char c )
 {
 	while ( *s ) *s++ = c ;
 }
+
